requirements.txt
asgiref==3.5.2
Django==4.1.3
django-axes==5.39.0
django-ipware==4.0.2
sqlparse==0.4.3
django-filter==22.1